Reorder Docker commands to reduce build time
This change should reduce the time needed to build a new docker image.
The premise is that the most expensive, most stable actions should be
first, to take advantage of cache. FreeSurfer release is out, and takes
longer than AFNI/FSL, so it's moved forward. Git/graphviz are moved
toward the end, as these are cheap, and make a natural place to add new
apt packages without forcing a re-fetch of FreeSurfer/AFNI/FSL/ANTs/etc.