ENH: Validate graymid inputs from FreeSurferSource
FreeSurferSource returns an OutputMultiPath, so the graymid input should
follow suit. One particular possibility is that a pre-computed subject
directory may have ?h.graymid in it and we create ?h.midthickness
(copies of ?h.graymid, since the introduction of MakeMidthickness).

A second run-through would receive both, which would cause an error.
This change accepts either as input, preferring ?h.midthickness when
available.

It also rejects files from the wrong hemisphere, which means this now
can be safely used from a FreeSurferSource that did not specify
hemisphere.