[ENH] Improve reading fieldmap-correction tags
This PR adds two new functions that can be used to read
fieldmap-correction related info from metadata.

Both are added with doctests that check they work in
various use-cases.

Some changes have been proposed to the BIDS specification
related to this issue. I'll submit an issue to INCF/BIDS-validator
with some heuristics that can be used to check for sanity of
echo spacing information in EPI scans.