[ENH] Change default template output to 2mm
This PR resurrects the dicussion about native or other output resolutions. In particular, this PR CHANGES the current default (the "native" space of the input BOLD file) and sets the template-2mm version.

Implications:

  * As off this PR, any new template added to FMRIPREP will require having a 1mm and a 2mm resolution versions. This was probably necessary already, but from this PR on it will be mandatory for sure.
  * The default behavior until v1.0.8 can be obtained using ``--output-grid-reference native``.
  * The option has been renamed: from ``--output-grid-reference`` to ``--template-resampling-grid``, as it is more descriptive.

Justification:

  * Comparison studies need not only a standard (physical) space, but also a standard voxel size (digital space) for the good of analysis.
  * Cluster analyses on images with very large voxels may be very hard to interpret.

Since it's been a long overdue request, I thought it was time to change this default. Closes #337.

Modified ds005 test on circleci with ``--template-resampling-grid native``, so that we test both options.