[FIX] Run ICA-AROMA in ``MNI152Lin`` space
This PR is an attempt to finally close #870, by ensuring that ICA-AROMA is run in the MNI152Lin space.

To do so, the corresponding workflow has been modified, including an internal BOLD resampling workflow. When the user specifies ``MNI152Lin`` as template space, this workflow will be duplicated. One possible optimization to consider in the future is identifying that situation, and avoid the outer resampling when AROMA is selected. However, for a first take I assumed this duplication is not too costly.

Other improvements included with this PR are:

  * [x] Update ICA-AROMA to the latest 0.4.4 release.
  * [x] Remove the check of ``--use-aroma`` and template.
  * [x] Update documentation accordingly.
  * [x] Update naming of non-aggresively denoised images to the new space