MAINT: Update PR template
As the guidelines are intended for the PR submitter, and not the reviewer, I added them to the comments, and changed the section name to "Pull request agreement".
Additionally, I made the final checklist item something everyone can check. This will avoid the "3 of 4" items complete indicator in the pull request list.