FIX: Generate proper LTA transform prior BOLD sampling on surfaces
This is a temporary patch before we go all the way in with NiTransforms
in the sampling of BOLD on surfaces.

The anatomical _fast-track_ required to expose the fsnative-to-T1w
transform in the derivatives folder (which we were already doing in ITK
format).

When fMRIPrep ran without the fast-track, then the LTA transform would
be directly passed in without conversions. The fast-track PR forced the
implementation to use the ITK version.

This, in conjunction with the little trick to stick the BOLD shape and
zooms into the LTA (i.e., using ``lta_concatenate`` with an identity
transform with those features, shape and zooms, as moving) resulted in
an overly complex workflow that I partially implemented with
NiTransforms.

This PR gets rid of the concatenation with identity trick, using
NiTransforms to generate a transform equivalent to the concatenated LTA
we used to generate before the fast-track was introduced.

Resolves: #2145
Assign: @mgxd
Milestone: 20.1.0
Related: #2118, #2041, #2121.