FIX: Simplify STC logic for too short BOLD series
After #2461 and #2468, the ``TooShort`` value for running STC could not
be assigned.
This PR ignores manually set dummy timepoints, although the case when
this could be problematic is largely a very edge case (say you have only
7 timepoints and manually set 3 as dummy).

On the flip side, this PR drops the use of a function discontinued in
the latest niworkflows release, fixing a bug ahead of time.