fix: revise @effigies' b67636b & e43a8fd
Turns out that our bold-to-bold resampling workflow has fake inputs and
outputs called ``bold_mask``.
That made me believe that, if we were joining the bold echoes from the
bold-to-bold resampling workflow, we could do the same with the mask.

Looking into this in depth made me realize of the confusing i/o spec of
the workflow, and also of the fact that it is unnecessary to join
anything in the absence of fieldwarp because the mask is the same you
calculated with the initial boldref.
So I connected that one and locally seems to be functioning okay.

Hopefully, this was the last bit of this not-so-trivial attempt to
organize the ME-EPI pathway. I'm already shaking by thinking at some
point we will need a similar in-depth revision of the SBRef dealings...